renamed template folder to 'dropdown_template'
