+++ plugins/dropdown_template/miradorTemplateDialog.js
