+++ plugins/dropdown_template/miradorTemplatePlugin.js
