+++ public/netlify.html
@@ -50,7 +50,7 @@
                     <iframe id="miradorViewer"
                         title="Mirador Viewer"
                         width="1200px"
-                        height="700px"
+                        height="100%"
                         src=""
                         allowfullscreen="true">
                     </iframe>
